Model-GCP/requirements.txt
numpy==1.26.4
pillow==11.0.0
tensorflow==2.16.1


